Fix ROS serial communication between Raspberry Pi and Arduino
Though there is still some issues with the feedback from the Arduino,
these issues are considered a lower priority, and the serial
communication package works adequately for GRO501-1. The issues MAY be
addessed for GRO501-2 validation.

NOTES:

- The changes in this commit have been tested (successfully) on the
  racecar with a Raspberry Pi 5 8GB running Ubuntu24.04;
- The requirements for GRO501-1 are respected (solutions for ROS1 were
  used for the validation: confirmed working appropriately);
- This commit does NOT contain any solution for GRO501-1: the solution
  code used for validating the serial communication and GRO501-1 were
  removed, and only the baseline ROS2 and low-level controller code are
  present.

CHANGES:

- Add `arduino_sensors` node to `teleop.launch.py` for functional
  teleoperation;
- Remove unnecessary, commented code from teleoperation launch file;
- Adjust serial communication values and logic;
- Add a code block for handling GRO501-1's "cruise control" scenario;
- Rename command and feedback topics to be more explicit;
- Remove compilation warnings from `PBUtils.cpp`;
- Add more explicit comments for GRO501-1 in `slash_teleop`.

+++ racecar_arduino/Controller/PBUtils.cpp
@@ -78,7 +78,7 @@ void PBUtils::pbSend(int nbs, ...)
       toSendBuilder += String(id);
       toSendBuilder += "|";
       
-      for(int j = 0; j < stream.bytes_written; j++)
+      for(size_t j = 0; j < stream.bytes_written; j++)
       {
         sprintf (toSend, "%02X", bufferOut[j]);
         toSendBuilder += String(toSend);
